backend/requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-cors==4.0.0
flask-jwt-extended==4.5.3
python-dotenv==1.0.0
psycopg2-binary==2.9.9
google-generativeai==0.3.1
python-docx==1.0.1
PyPDF2==3.0.1
bcrypt==4.0.1
gunicorn==21.2.0
